Encapsulate working with sources
