+++ plugin-testlib/src/main/java/io/spine/tools/gradle/testing/GradleProject.java
@@ -26,22 +26,12 @@
 
 package io.spine.tools.gradle.testing;
 
-import com.google.common.collect.Multimap;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
-import io.spine.tools.gradle.SourceSetName;
 import io.spine.tools.gradle.task.TaskName;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Path;
-import java.util.Map;
-
-import static com.google.common.base.Preconditions.checkState;
-import static java.lang.String.format;
-import static java.nio.file.Files.copy;
-import static java.nio.file.Files.createDirectories;
 
 /**
  * Allows to configure a Gradle project for testing needs.
@@ -49,9 +39,6 @@ import static java.nio.file.Files.createDirectories;
  * <p>The project operates in the given test project directory and allows executing Gradle tasks.
  */
 public final class GradleProject {
-
-    /** The name of the directory under {@code resources} which contains files for this project. */
-    private final String origin;
 
     private final GradleRunner runner;
     private final RunnerArguments arguments;
@@ -70,22 +57,23 @@ public final class GradleProject {
         return "java";
     }
 
-    GradleProject(GradleProjectSetup builder) throws IOException {
-        this.origin = builder.origin();
-        this.arguments = builder.arguments();
+    GradleProject(GradleProjectSetup setup) throws IOException {
+        this.arguments = setup.arguments();
         this.runner = GradleRunner.create()
-                .withProjectDir(builder.dir())
-                .withDebug(builder.debug());
-        if (builder.addPluginUnderTestClasspath()) {
+                .withProjectDir(setup.dir())
+                .withDebug(setup.debug());
+        if (setup.addPluginUnderTestClasspath()) {
             runner.withPluginClasspath();
         }
-        if (builder.environment() != null) {
-            runner.withEnvironment(builder.environment());
+        if (setup.environment() != null) {
+            runner.withEnvironment(setup.environment());
         }
-        writeGradleScripts();
-        writeBuildSrc();
-        writeProtoFiles(builder.protoFileNames());
-        writeJavaFiles(builder.javaFileNames());
+        writeSources(setup);
+    }
+
+    private static void writeSources(GradleProjectSetup setup) throws IOException {
+        Sources sources = new Sources(setup);
+        sources.write();
     }
 
     /**
@@ -94,38 +82,6 @@ public final class GradleProject {
     @SuppressWarnings("unused")
     public GradleRunner runner() {
         return runner;
-    }
-
-    private void writeGradleScripts() throws IOException {
-        Path projectDir = projectDir();
-        BuildGradle buildGradle = new BuildGradle(projectDir);
-        buildGradle.createFile();
-
-        TestEnvGradle testEnvGradle = new TestEnvGradle(projectDir);
-        testEnvGradle.createFile();
-    }
-
-    /** The directory of this project. */
-    private Path projectDir() {
-        return runner.getProjectDir()
-                     .toPath();
-    }
-
-    private void writeBuildSrc() throws IOException {
-        Path targetDir = projectDir();
-        BuildSrc.writeTo(targetDir);
-    }
-
-    private void writeProtoFiles(Multimap<SourceSetName, String> fileNames) throws IOException {
-        for (Map.Entry<SourceSetName, String> protoFile : fileNames.entries()) {
-            writeProto(protoFile.getKey(), protoFile.getValue());
-        }
-    }
-
-    private void writeJavaFiles(Multimap<SourceSetName, String> fileNames) throws IOException {
-        for (Map.Entry<SourceSetName, String> javaFile : fileNames.entries()) {
-            writeJava(javaFile.getKey(), javaFile.getValue());
-        }
     }
 
     @CanIgnoreReturnValue
@@ -141,38 +97,5 @@ public final class GradleProject {
     private GradleRunner prepareRun(TaskName taskName) {
         String[] args = arguments.forTask(taskName);
         return runner.withArguments(args);
-    }
-
-    private void writeProto(SourceSetName ssn, String fileName) throws IOException {
-        writeFile(fileName, protoDir(ssn));
-    }
-
-    static String protoDir(SourceSetName ssn) {
-        return format("src/%s/proto/", ssn);
-    }
-
-    private void writeJava(SourceSetName ssn, String fileName) throws IOException {
-        writeFile(fileName, javaDir(ssn));
-    }
-
-    private static String javaDir(SourceSetName ssn) {
-        return format("src/%s/java/", ssn);
-    }
-
-    private void writeFile(String fileName, String dir) throws IOException {
-        String filePath = dir + fileName;
-        String resourcePath = origin + '/' + filePath;
-        try (InputStream fileContent = openResource(resourcePath)) {
-            Path fileSystemPath = projectDir().resolve(filePath);
-            createDirectories(fileSystemPath.getParent());
-            copy(fileContent, fileSystemPath);
-        }
-    }
-
-    private InputStream openResource(String fullPath) {
-        InputStream stream = getClass().getClassLoader()
-                                       .getResourceAsStream(fullPath);
-        checkState(stream != null, "Unable to locate resource: `%s`.", fullPath);
-        return stream;
     }
 }
